requirements.txt
colorama==0.4.6
termcolor==2.3.0
Pillow==10.0.0
customtkinter>=5.2.0
PyQt6>=6.4.0
tkinterweb>=4.0.6
packaging>=23.0 